+++ objects/blueguy.h
@@ -2,7 +2,7 @@
 #define OBJ_BLUEGUY_H
 
 int blueguyInit();
-int blueguyNew(int x, int y);
+int blueguyNew(float x, float y);
 int blueguyDraw();
 
 #endif
